backend/requirements.txt
fastapi
uvicorn
azure-data-tables
azure-storage-blob
pandas
openpyxl
python-multipart
python-dotenv
pydantic-settings
passlib[sha256_crypt]
python-jose[cryptography]
gunicorn
